Add support for getting database defaults for objects to the API.
Go has strict ideas about what variables should default to and variable
nullibility, and thus the Go API needs a way to figure out what the
default values should be for any given object type we expose via the
API.  To make this happen, and to make life simpler for other future API
clients, all of the controllers have been updated to allow clients to
get a sample object of whatever type they manage.  This sample is
populated with the database column defaults for the attributes it
exposes.

While we are at it, fix up some JSON serialization and overexposure in
the API.