Add support for a non-secret-leaking password change protocol.
This allows us to change the password for a user via the API
without ever transmitting the password in cleartext, even over an
unsecure protocol. The password change protocol looks like this:

Client: GET api/v2/users/userid/start_password_reset -> Server

Server checks to see if the current user can change the password of
userid.  If not, server sleeps for 2 seconds then raises an
exception. Otherwise, Server deletes any other change tokens issued for
User, generates a one time use public/private keypair, saves both
halves, and returns the base64 encoded public key to Client.

Client generates its own public/private keypair and a random nonce. It
then encrypts ```{"password":"new_password"}``` using its private key,
the Server's generated public key, and the nonce, yielding the payload.

Client: POST api/v2/users/userid/complete_password_reset -> Server

The body of the POST is a JSON document with the following format:

    {
      "token":"base64 encoded serverPublicKey",
      "decoder":"base64 encoded clientPublicKey",
      "nonce":"base64 encoded client-generated random nonce",
      "payload":"base64 encoded encrypted password JSON blob",
      "digest":true
    }

The server looks up the change token based on the "token" field, and
verifies that the current user can change the password, that the token
has not expired, and that the requested token to use is associated with
the user whose password is to be changed.  If not, the server sleeps for
2 seconds then raises an exception. Otherwise, the server decodes the
payload using the token's private key, the passed-in client public key,
and the nonce. It then throws away the token, and saves the new password
in a digest authentication compatible format.

This protocol uses the nacl cryptosystem throughout for ease of
implementation.